# More efficient encoding of floats and doubles without exponent