Tested with Scala 3 and GraalVM JDK-25ea on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                            (size)   Mode  Cnt        Score        Error  Units
ArrayOfDoublesWriting.zioJson           128  thrpt    5   213042.974 ±   1967.557  ops/s
ArrayOfDoublesWriting.zioSchemaJson     128  thrpt    5   210949.814 ±   2905.848  ops/s
ArrayOfFloatsWriting.zioJson            128  thrpt    5   285297.646 ±   7687.284  ops/s
ArrayOfFloatsWriting.zioSchemaJson      128  thrpt    5   314495.765 ±   5640.222  ops/s
GeoJSONWriting.zioJson                  N/A  thrpt    5    16847.105 ±   1045.649  ops/s
GeoJSONWriting.zioSchemaJson            N/A  thrpt    5    26026.200 ±    349.518  ops/s
```

#### After
```txt
Benchmark                            (size)   Mode  Cnt        Score        Error  Units
ArrayOfDoublesWriting.zioJson           128  thrpt    5   216015.004 ±   5348.566  ops/s
ArrayOfDoublesWriting.zioSchemaJson     128  thrpt    5   220146.565 ±   1437.523  ops/s
ArrayOfFloatsWriting.zioJson            128  thrpt    5   337490.570 ±  22496.827  ops/s
ArrayOfFloatsWriting.zioSchemaJson      128  thrpt    5   331373.715 ±   3244.506  ops/s
GeoJSONWriting.zioJson                  N/A  thrpt    5    14534.595 ±   2277.668  ops/s
GeoJSONWriting.zioSchemaJson            N/A  thrpt    5    32090.535 ±    682.181  ops/s
```